ShowMoreButton rework
