+++ osu.Game/Overlays/Comments/Buttons/ShowMoreButton.cs
@@ -8,21 +8,28 @@ using osu.Framework.Graphics.Containers;
 using osu.Game.Graphics.Sprites;
 using System.Collections.Generic;
 using osuTK;
+using osu.Framework.Allocation;
 
-namespace osu.Game.Overlays.Comments
+namespace osu.Game.Overlays.Comments.Buttons
 {
-    public abstract class GetCommentRepliesButton : LoadingButton
+    public class ShowMoreButton : LoadingButton
     {
-        private const int duration = 200;
-
         protected override IEnumerable<Drawable> EffectTargets => new[] { text };
 
         private OsuSpriteText text;
 
-        protected GetCommentRepliesButton()
+        public ShowMoreButton()
         {
             AutoSizeAxes = Axes.Both;
+            Margin = new MarginPadding { Vertical = 10, Left = 80 };
             LoadingAnimationSize = new Vector2(8);
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(OverlayColourProvider colourProvider)
+        {
+            IdleColour = colourProvider.Light2;
+            HoverColour = colourProvider.Light1;
         }
 
         protected override Drawable CreateContent() => new Container
@@ -31,15 +38,13 @@ namespace osu.Game.Overlays.Comments
             Child = text = new OsuSpriteText
             {
                 AlwaysPresent = true,
-                Font = OsuFont.GetFont(size: 12, weight: FontWeight.Bold),
-                Text = GetText()
+                Font = OsuFont.GetFont(size: 12, weight: FontWeight.SemiBold),
+                Text = "show more"
             }
         };
 
-        protected abstract string GetText();
+        protected override void OnLoadStarted() => text.FadeOut(200, Easing.OutQuint);
 
-        protected override void OnLoadStarted() => text.FadeOut(duration, Easing.OutQuint);
-
-        protected override void OnLoadFinished() => text.FadeIn(duration, Easing.OutQuint);
+        protected override void OnLoadFinished() => text.FadeIn(200, Easing.OutQuint);
     }
 }
